requirements.txt
Django==3.0.1
djangorestframework==3.13.1
geopy==2.2.0
geotiff==0.2.4
gunicorn==20.1.0
mysql==0.0.3
mysql-connector-python==8.0.28
mysqlclient==2.1.0
Pillow==8.4.0
plotly==5.6.0